--HG-- extra : convert_revision : svn%3A946a0da7-ebce-4904-9acb-2f1e67aed693%4038

+++ Steam3Lib/Util/DataStream.cs
@@ -258,9 +258,9 @@ namespace SteamLib
         /// Sizes the remaining size.
         /// </summary>
         /// <returns></returns>
-        public long SizeRemaining()
+        public uint SizeRemaining()
         {
-            return Length - Position;
+            return ( uint )Length - ( uint )Position;
         }
